remove utils to utils/weight_init
