+++ libai/models/bert_model.py
@@ -14,6 +14,8 @@
 # limitations under the License.
 
 import oneflow as flow
+from oneflow import nn
+
 from libai.layers import (
     build_activation,
     VocabEmbedding,
@@ -25,10 +27,8 @@ from libai.layers import (
     LMLogits,
 )
 from libai.utils import distributed as dist
-from oneflow import nn
 from libai.config import configurable
-
-from .utils import init_method_normal, scaled_init_method_normal
+from libai.models.utils.weight_init import init_method_normal, scaled_init_method_normal
 
 from .build import MODEL_ARCH_REGISTRY
